+++ locust/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.0.0b0"
+__version__ = "2.0.0b1"
 
 from gevent import monkey