+++ locust/test/test_main.py
@@ -14,7 +14,7 @@ from locust import main
 from locust.argument_parser import parse_options
 from locust.main import create_environment
 from locust.user import HttpUser, User, TaskSet
-from .mock_locustfile import mock_locustfile, MOCK_LOUCSTFILE_CONTENT
+from .mock_locustfile import mock_locustfile, MOCK_LOCUSTFILE_CONTENT
 from .testcases import LocustTestCase
 from .util import temporary_file, get_free_tcp_port
 
@@ -71,7 +71,7 @@ class TestLoadLocustfile(LocustTestCase):
 
     def test_with_shape_class(self):
         content = (
-            MOCK_LOUCSTFILE_CONTENT
+            MOCK_LOCUSTFILE_CONTENT
             + """class LoadTestShape(LoadTestShape):
     pass
         """
@@ -231,14 +231,25 @@ class LocustProcessIntegrationTest(TestCase):
         with mock_locustfile() as mocked:
             output = (
                 subprocess.check_output(
-                    ["locust", "-f", mocked.file_path, "--host", "https://test.com/", "--run-time", "1s", "--headless"],
+                    [
+                        "locust",
+                        "-f",
+                        mocked.file_path,
+                        "--host",
+                        "https://test.com/",
+                        "--run-time",
+                        "1s",
+                        "--headless",
+                        "--loglevel",
+                        "DEBUG",
+                    ],
                     stderr=subprocess.STDOUT,
                     timeout=2,
                 )
                 .decode("utf-8")
                 .strip()
             )
-            self.assertIn("Spawning 1 users at the rate 1 users/s", output)
+            self.assertIn('Spawning additional {"UserSubclass": 1} ({"UserSubclass": 0} already running)...', output)
 
     def test_headless_spawn_options_wo_run_time(self):
         with mock_locustfile() as mocked:
@@ -257,13 +268,13 @@ class LocustProcessIntegrationTest(TestCase):
             self.assertIn("Shutting down (exit code 0), bye", stderr)
 
     def test_default_headless_spawn_options_with_shape(self):
-        content = MOCK_LOUCSTFILE_CONTENT + textwrap.dedent(
+        content = MOCK_LOCUSTFILE_CONTENT + textwrap.dedent(
             """
             class LoadTestShape(LoadTestShape):
                 def tick(self):
                     run_time = self.get_run_time()
                     if run_time < 2:
-                            return (10, 1)
+                        return (10, 1)
 
                     return None
             """
@@ -340,9 +351,11 @@ class LocustProcessIntegrationTest(TestCase):
                         mocked.file_path,
                         "--headless",
                         "--run-time",
-                        "4s",
+                        "7s",
                         "-u",
                         "0",
+                        "--loglevel",
+                        "INFO",
                     ]
                 ),
                 stderr=STDOUT,
@@ -353,22 +366,31 @@ class LocustProcessIntegrationTest(TestCase):
             gevent.sleep(1)
 
             stdin.write(b"w")
-            gevent.sleep(0.1)
+            gevent.sleep(1)
             stdin.write(b"W")
-            gevent.sleep(0.1)
+            gevent.sleep(1)
             stdin.write(b"s")
-            gevent.sleep(0.1)
+            gevent.sleep(1)
             stdin.write(b"S")
+            gevent.sleep(1)
 
+            # This should not do anything since we are already at zero users
+            stdin.write(b"S")
             gevent.sleep(1)
 
             output = proc.communicate()[0].decode("utf-8")
             stdin.close()
-            self.assertIn("Spawning 1 users at the rate 100 users/s", output)
-            self.assertIn("Spawning 10 users at the rate 100 users/s", output)
-            self.assertIn("1 Users have been stopped", output)
-            self.assertIn("10 Users have been stopped", output)
+            self.assertIn("Ramping to 1 users using a 100.00 spawn rate", output)
+            self.assertIn('All users spawned: {"UserSubclass": 1} (1 total users)', output)
+            self.assertIn("Ramping to 11 users using a 100.00 spawn rate", output)
+            self.assertIn('All users spawned: {"UserSubclass": 11} (11 total users)', output)
+            self.assertIn("Ramping to 10 users using a 100.00 spawn rate", output)
+            self.assertIn('All users spawned: {"UserSubclass": 10} (10 total users)', output)
+            self.assertIn("Ramping to 0 users using a 100.00 spawn rate", output)
+            self.assertIn('All users spawned: {"UserSubclass": 0} (0 total users)', output)
             self.assertIn("Test task is running", output)
+            self.assertIn("Shutting down (exit code 0), bye.", output)
+            self.assertEqual(0, proc.returncode)
 
     def test_html_report_option(self):
         with mock_locustfile() as mocked: