+++ docs/changelog.rst
@@ -4,6 +4,23 @@ Changelog Highlights
 
 For full details of the Locust changelog, please see https://github.com/locustio/locust/blob/master/CHANGELOG.md
 
+2.0.0
+=====
+
+* Auto shrink request stats table to fit terminal https://github.com/locustio/locust/pull/1811 (2.0.0b2)
+* Refactoring of the dispatch logic to improve performance https://github.com/locustio/locust/pull/1809 (2.0.0b2)
+* Check version of workers when they connect. Warn if there is a mismatch, refuse 1.x workers to connect https://github.com/locustio/locust/pull/1805 (2.0.0b1)
+* Change the default User weight to 1 instead of 10 https://github.com/locustio/locust/pull/1803 (2.0.0b1)
+* Upgrade to flask 2 https://github.com/locustio/locust/pull/1764 (2.0.0b1)
+* Move User selection responsibility from worker to master in order to fix unbalanced distribution of users and uneven ramp-up https://github.com/locustio/locust/pull/1621 (2.0.0b0)
+
+1.6.0
+=====
+
+* Allow cross process communication using custom messages https://github.com/locustio/locust/pull/1782
+* Fix: status "stopped" instead of "spawning", tick\(\) method of LoadShape called only once https://github.com/locustio/locust/pull/1769
+* Other small fixes (see https://github.com/locustio/locust/blob/master/CHANGELOG.md)
+
 1.5.3
 =====